+++ PageManagements/Assets/PageManagements/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [1.0.1] - 2024-07-11
 ### Added
 - Added samples.
-- Added HasPage() into manager.
+- Added GetCount() and HasPage() into manager.
+- Added PageChanged event into manager.
 
 ## [1.0.0] - 2024-07-10
 ### Added